Fix: Registration flow, error messaging and handling
